Module Mirage_block_combinators.Make_safe

Safe block devices

Construct a safe wrapper around B where necessary buffer preconditions are checked on read and write, and useful error messages generated. Some concrete implementations generate confusing errors (e.g. Unix might say "EINVAL") which are harder to debug.

Parameters

B : Mirage_block.S

Signature

type error#row

The type for errors.

and error = private [>

| Mirage_block.error

| `Unsafe of string

]

The type for errors.

type write_error#row

The type for write errors.

and write_error = private [>

| Mirage_block.write_error

| `Unsafe of string

]

The type for write errors.

include Mirage_block.S with type S.t = B.t and type error := error and type write_error := write_error

val pp_error : error Fmt.t

val pp_write_error : write_error Fmt.t

type t = B.t

val disconnect : t -> unit Lwt.t

val get_info : t -> Mirage_block.info Lwt.t

val read : t -> int64 -> Cstruct.t list -> (unit, error) Stdlib.result Lwt.t

val write : t -> int64 -> Cstruct.t list -> (unit, write_error) Stdlib.result Lwt.t

val unsafe_read : t -> int64 -> Cstruct.t list -> (unit, B.error) Stdlib.result Lwt.t

unsafe_read is like read except it bypasses the necessary buffer precondition checks. Only use this if you want maximum performance and if you can prove the preconditions are respected.

val unsafe_write : t -> int64 -> Cstruct.t list -> (unit, B.write_error) Stdlib.result Lwt.t

unsafe_write is like write except it bypasses the necessary buffer precondition checks. Only use this if you want maximum performance and if you can prove the buffer preconditions are respected.